feat: Add image delete API and refactor server code
- Add DELETE /api/v0/images/{imageID} endpoint
- Users can only delete their own images (authenticated)
- Move S3 client to new internal/clients package
- Extract image/presign/status/delete handlers into separate files
- Add image ID (S3 UUID) to GET /api/v0/images response
- Update API documentation with new endpoint and id field
- Implement ownership validation at S3 key level (userID/imageID)
